Refactored filterCriterionToUrlParam into CudamiRestClient, so that it can be used by CudamiTagsClient and CudamiSubjectsClient
