+++ dc-cudami-client/src/main/java/de/digitalcollections/cudami/client/CudamiRestClient.java
@@ -5,11 +5,17 @@ import de.digitalcollections.client.BaseRestClient;
 import de.digitalcollections.model.UniqueObject;
 import de.digitalcollections.model.exception.TechnicalException;
 import de.digitalcollections.model.exception.http.client.ResourceNotFoundException;
+import de.digitalcollections.model.list.filtering.FilterCriterion;
+import de.digitalcollections.model.list.filtering.FilterOperation;
 import de.digitalcollections.model.list.paging.PageRequest;
 import de.digitalcollections.model.list.paging.PageResponse;
+import java.net.URLEncoder;
 import java.net.http.HttpClient;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.UUID;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -61,5 +67,38 @@ public class CudamiRestClient<T extends UniqueObject> extends BaseRestClient<T> 
 
   public T update(UUID uuid, T object) throws TechnicalException {
     return doPutRequestForObject(String.format("%s/%s", baseEndpoint, uuid), object);
+  }
+
+  @Override
+  protected String filterCriterionToUrlParam(FilterCriterion filterCriterion) {
+    Matcher labelOrName = Pattern.compile("^(label|name)").matcher(filterCriterion.getExpression());
+    if (labelOrName.find()) {
+      if (filterCriterion.getValue() == null || !(filterCriterion.getValue() instanceof String)) {
+        return "";
+      }
+      String value = (String) filterCriterion.getValue();
+      if (filterCriterion.getOperation() == FilterOperation.EQUALS) {
+        value = String.format("\"%s\"", value);
+      } else if (filterCriterion.getOperation() != FilterOperation.CONTAINS) {
+        throw new UnsupportedOperationException(
+            "`label` and `name` can only be filtered by using CONTAINS (should be preferred) or EQUALS!");
+      }
+      String urlParams =
+          String.format(
+              "%s=%s", labelOrName.group(1), URLEncoder.encode(value, StandardCharsets.UTF_8));
+      Matcher matchLanguage =
+          Pattern.compile("\\.([\\w_-]+)$").matcher(filterCriterion.getExpression());
+      if (matchLanguage.find()) {
+        // there is a language defined
+        return urlParams
+            + String.format(
+                "&%sLanguage=%s",
+                labelOrName.group(1),
+                URLEncoder.encode(matchLanguage.group(1), StandardCharsets.UTF_8));
+      }
+      return urlParams;
+    }
+
+    return super.filterCriterionToUrlParam(filterCriterion);
   }
 }
