+++ dc-cudami-client/src/main/java/de/digitalcollections/cudami/client/identifiable/CudamiIdentifiablesClient.java
@@ -6,23 +6,18 @@ import de.digitalcollections.model.exception.TechnicalException;
 import de.digitalcollections.model.exception.http.client.ResourceNotFoundException;
 import de.digitalcollections.model.identifiable.Identifiable;
 import de.digitalcollections.model.identifiable.alias.LocalizedUrlAliases;
-import de.digitalcollections.model.list.filtering.FilterCriterion;
-import de.digitalcollections.model.list.filtering.FilterOperation;
 import de.digitalcollections.model.list.paging.PageRequest;
 import de.digitalcollections.model.list.paging.PageResponse;
 import de.digitalcollections.model.list.sorting.Direction;
 import de.digitalcollections.model.list.sorting.NullHandling;
 import de.digitalcollections.model.list.sorting.Order;
 import de.digitalcollections.model.list.sorting.Sorting;
-import java.net.URLEncoder;
 import java.net.http.HttpClient;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.UUID;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import org.apache.commons.codec.binary.Base64;
 
@@ -44,39 +39,6 @@ public class CudamiIdentifiablesClient<I extends Identifiable> extends CudamiRes
         (Class<I>) Identifiable.class,
         mapper,
         API_VERSION_PREFIX + "/identifiables");
-  }
-
-  @Override
-  protected String filterCriterionToUrlParam(FilterCriterion filterCriterion) {
-    Matcher labelOrName = Pattern.compile("^(label|name)").matcher(filterCriterion.getExpression());
-    if (labelOrName.find()) {
-      if (filterCriterion.getValue() == null || !(filterCriterion.getValue() instanceof String)) {
-        return "";
-      }
-      String value = (String) filterCriterion.getValue();
-      if (filterCriterion.getOperation() == FilterOperation.EQUALS) {
-        value = String.format("\"%s\"", value);
-      } else if (filterCriterion.getOperation() != FilterOperation.CONTAINS) {
-        throw new UnsupportedOperationException(
-            "`label` and `name` can only be filtered by using CONTAINS (should be preferred) or EQUALS!");
-      }
-      String urlParams =
-          String.format(
-              "%s=%s", labelOrName.group(1), URLEncoder.encode(value, StandardCharsets.UTF_8));
-      Matcher matchLanguage =
-          Pattern.compile("\\.([\\w_-]+)$").matcher(filterCriterion.getExpression());
-      if (matchLanguage.find()) {
-        // there is a language defined
-        return urlParams
-            + String.format(
-                "&%sLanguage=%s",
-                labelOrName.group(1),
-                URLEncoder.encode(matchLanguage.group(1), StandardCharsets.UTF_8));
-      }
-      return urlParams;
-    }
-
-    return super.filterCriterionToUrlParam(filterCriterion);
   }
 
   public List<I> find(String searchTerm, int maxResults) throws TechnicalException {
